refactor(state): share transcript path normalization
